test/spanner/3_sample/requirements.txt
google-cloud-monitoring==2.19.0
